requirements.txt
qrcode
tqdm
colorama
pillow